test(CI and manual): changes the default test browser to ChromeHeadless
The tests previously defaulted to using PhantomJS for headless testing.
PhantomJS is now deprecated and produces unwanted warnings in the test output.
The default test browser changes to ChromeHeadless.

GitHub Actions are added to allow testing on Ubuntu, MacOS and Windows Server.
Tests are run in ChromeHeadless and FirefoxHeadless.
This gives better test coverage across the latest browsers and operating systems.