test(index.spec.js): fixes a floating point issue in a test
This fixes a floating point issue with a test.
The original test passes in PhantomJs but fails in Chrome and Firefox.
This is presumably due to some internal difference in floating point calculations.

The algorithm used is copied from the Jest expect library which has a built-in
expectation matcher toBeCloseTo(number, numDigits?).
This is not in the version of expect we currently use.