test(index.spec.js): fixes first run timeout error by increasing the timeout from 2 to 5 seconds
When the tests first run they can fail due to a timeout error.
The default timeout for the tests is 2000ms.

The delays occur because the images have to be obtained from unsplash.com.
The first fetch of an image takes about 1 second for me.
Hence, if two new images have to be fetched within a single test then the
default timeout can be exceeded.

The timeout for all tests is set to 5 seconds.

Note that setting the timeout requires the Mocha context, hence the change
from an arrow function to a normal function within the "describe" block.
(Mocha recommends not using arrow functions for this reason.)