fix(InnerImageZoom.js): cleans up the zoom out timer on unmount and zoom in
This fix removes warnings from the unit test runs.
The warnings are produced when a timeout function executes after
the component has unmounted.
The timer is now cleared when the component unmounts.

In normal use these errors would probably not be noticed because the
timeout duration is relatively short compared to the time between user clicks.

If the fadeDuration config is set to a long value then debouncing of the
zoom in and zoom out is required. For this reason the timer is also
cleared on zoom in.

Other potential zoom in/out debounce issues could still be present
because the zoom out timer also resets several variables.