fix(InnerImageZoom): Fix a bug in Firefox when the cursor is dragged to the left side of the window.
This bug affects Firefox when zoomed images are dragged.

When the cursor is dragged to the left side of the browser window the e.pageX value can be 0.
This zero x offset is interpreted as a falsy value by the code and e.changedTouches[0].pageX
is executed instead. This causes an exception because e.changedTouches is not defined.

This change allows an x offset of 0 to be recognised as a valid value in function handleDragMove.

The same fix is applied to function handleDragStart for consistency.