fix(InnerImageZoom): Remove unwanted whitespace from className.
The template string that combines multiple classes leads to unwanted whitespace
at the end of the class string.

The extra whitespace does not cause a problem with HTML parsing since extra
whitespace is ignored. It does look odd in the rendered code however, and
it is slightly more efficient to only include the minimum whitespace required.
Other solutions are available, such as the Node modules classnames and clsx,
but these would add a module dependency that is not required.

The template string has been adjusted such that the separator character
(a single space) is now only added when an optional class is added.