fix(InnerImageZoom): Add width unit (px) to optional inline style value
The width property is optional. When present it is specified to be a number.
Allow for the width property being undefined.

At the moment it looks like the width can only be specified in pixels.
If the width is specified as a string, rather than a number, it will be ignored.