fix(InnerImageZoom): Cancels the fade out when the mouse enters.
This applies to image panning.

When the image is zoomed, if the mouse is moved out and then back in quickly
an invalid state is entered and the image no longer zooms when clicked.
Zooming can only be re-enabled by mousing out and back in for a second time.

This happens because handleMouseEnter() executes before handleFadeOut(),
the opposite to the expected order when the mouse movement is slower.

The frequency of this bug occurring would be greater as the fadeDuration
property is increased. The default is 150ms.